Deploy: Increase swap size for Raspberry Pi builds
Deploy: Increase swap size for Raspberry Pi builds

Raspbian sets the swap size to 100MB by default, which causes the npm install step to hang the system on devices with 500MB of RAM. About 700MB of total memory is needed for the build/install process.

This commit adds a step to build_rpanion.sh to increase the swap size, which enables the installation to complete successfully.

Fixes issue #207 and maybe #208 also